corrections, update dependencies - bug fixes

+++ examples/fib.c
@@ -24,6 +24,6 @@ int main(int argc, char *argv[])
      n = atoi(argv[1]);
      result = fib(n);
 
-     printf("Result: %zu\n", (unsigned long long)result);
+     printf("Result: %zu\n", (size_t)result);
      return 0;
 }
